Reduce unnecessary requires
Inside `app` everything is auto/eager-loaded by Rails, so you never need
to require anything else that lives inside `app`.

The way `lib` is invoked most typically is via `autoload_path` or in
pure Ruby, `ruby -I ./lib` or the `RUBYLIB` ENV var. So components
should be available to each other if naming conventions are followed.

We don't actually follow naming conventions fully (`Checksum` is not in
an `Audit` module), *but* for referencing peers in the same directory
that doesn't matter.