EndpointType.endpoint_class should be an Enum
for now, it is a string based enum.  as per the comments, it should eventually become a more standard int enum.  this should work fow now, will file a follow on ticket if this gets merged, as making it an int in the DB is less urgent than making it an Enum in the first place.