Update rubocop exceptions, and solve the each_key/each_value problem
The rule was triggering on non-hash `Config::Options` objects, because
they are very `Hash`-like.  Easy solution is to convert the part that we
care about to a regular hash, when we are going to iterate in a way that
only wants keys or only wants values.

This commit includes removing other exceptions that are no longer applicable!