Convert Profiler invocations in Audit module
This is a setup for externalizing each `xxx_profiled` method to the
place where it is called (seems like each is called in only one place).

After all, it shouldn't be the business logic for a class how a given
mehtod is going to be profiled (or which are profilable).