Replace noramlize_druid_name with more accurate validation
This method existed to correct previously inserted data (druids w/
prefixes), but it has served its purpose and is no longer needed.  The
more comprehensive validation ensures that no more druid prefixes can be
inserted (i.e. the method will never be needed again).

This also helps because one of the tests for the removed method was
silently invalid: was catching generic `raise_error`, but the actual
exception was coming from the mock, not the model!

Fixes #893