Endpoint Check Job with existence logic to endpoint
EndpointCheckJob#perform has some of the desired logic. If the object
exists in the endpoint then it will update the last_checksum_validation
attribute and print out a message. If it doesn't exist that is a TODO

compare stored checksums to the replicated s3_object checksum

Update preserved copy status with the given result