Convert "large" CV jobs to async basis
We dispense with the old ordering here in order to use the conventional
`.find_each`.  Note: order should not be a part of a scope in any case.

Async CV performance should be good enough and flexible enough that we
are less concerned about which row gets processed "first", since all of
them will be processed "soon enough".  Alternatively, we can circle back
with a multi-query approach that first enumerates the distinct:
`DATE_TRUNC('day', last_checksum_validation) AS last_cv ORDER BY last_cv`

for expired checksums and then retrieves a batch for each day. But that
can wait.