Add DVZ logic for #parts and #expected_parts
Because we will only be able to store the **count** of parts in the S3
metadata, we need a way of reconstructing what filename is expected to
correspond to part X of N, or the whole set of N filenames, as here.

This logic will be utilized after workers are converted subsequent to
the PC model split.