fix ArchiveEndpoint#which_need_archive_copy scope (#899)
or at least, fix the bug we know about.

the issue was that the subquery (as previously written) wasn't filtering preserved copies by druid, it was just checking to see whether an archive endpoint had *any* preserved copy of the specified version, regardless of the preserved object it belonged to.  the direct port of that scope from Endpoint to ArchiveEndpoint kept that bug.

this explains the behavior @SaravShah and @atz saw, where PreservedObject#create_archive_preserved_copies worked the very first time it was called, and then not again when invoked on a different druid's PreservedObject but the same version.

wrote a new scope, ArchiveEndpoint#which_have_archive_copy(druid, version), and used that as the subquery for the re-written #which_need_archive_copy scope.

refactored the test a bit to add an expectation that creating an archive entry for druid abc version 1 on endpoint east1 will have no effect on the query results for whether endpoint east1 needs a copy of version 1 of druid def.