Remove redundant modeling
The `has_one... through` is a more powerful version of the delegation
statement.

Also, no need to add additional validation for enum beyond presence,
since that is literally how enums already work.

For example, if we try to set invalid values, we are already blocked:
```ruby
> x = ZippedMoabVersion.new(status: 'xx')
ArgumentError: 'xx' is not a valid status

> x = ZippedMoabVersion.new
=> #<ZippedMoabVersion:0x007fa37c9e8b68 ...>
> x.status='qq'
ArgumentError: 'qq' is not a valid status
```

So it is not possible to get the object into a state where the
additional validation would usefully apply.