Handle blocking ZMV data around `drop_endpoint`
The value of the method is uncertain since every PC will have a ZMV row
(or several) because of the `after_create` hook, meaning `drop_endpoint`
is only useful if you are already going underneath it to clean up data.

This doesn't resolve the question, but keeps the test preconditions
equivalent to their previous expectations, so the tests pass.