Stop nuking the DB every time rspec runs
This practice seems to have derived from having non-rails_helper'd code
writing to the database, as with our old `./lib` directory.  It is
unnecessary for anything using rails_helper, however, which now includes all
the code that writes to the DB during tests.

We already have transactional fixtures turned on so every change
during the rspec suite is atomically rolled back at the end of the example
where it was made.

Also, we use the fixtures directory, but we don't use it for
ActiveRecord fixtures, so that config in rails_helper is misleading.