Add zip cache cleanup cronjob and capistrano role
Currently, this would run every hour and deletes any files older than 1 day.
So the oldest a file would be is 1 day (and 59 minutes). This may
require further tuning depending on how we choose to enqueue backlog for
`ZipmakerJob` or when we have completed the backlog and have different
operational expectations for volume of work.

Closes #1032