Mainly changes to Capistrano deploy docs
Our section on "how to deploy" did not actually describe how to deploy,
instead restating the lines from earlier in the file about DB seeding.

I don't see anything related to model splitting or renaming that needs
to be touched here.

Changed `rake spec` to `rspec` for concision and clarity.