Purge HostSettings references and unneeded or broken rake tasks
Our expecations about dropping and seeding roots has changed, especially
since we don't expect systems to only mount portions of roots, since
using async jobs means that a worker has to be prepared for a job on any
root.

In most places, HostSettings were being used to enumerate settings for
stuff that was already in the database anyway.  I think much of that
comes from before we had a `MoabStorageRoot`.  Now we know that a
location or value is good because it is one of the values in that table.
The DB is shared between many machines, so we won't be frequently adding
or dropping roots.