Change cleanup logic of sdr-transfers
Problem: We have large zip files that may be old(ish), awaiting md5 creation and shipping to the cloud. By basing our deletion criteria on the age of the zip, rather than the age of md5 sidecar files, we increase the risk that we'll delete zips that are either awaiting md5 creation or are still in the process of being transferred to endpoints.
Solution: Look at the creation time of MD5 sidecar files instead; if an md5 sidecar has existed for > 1 day, chances are better that the associated zip file has been successfully uploaded.