Create missing preservationAuditWF
Fixes #1258

The implementation was never matching the exception message due to unescaped parentheses in the regex. Fix that, and change the test to make it fail without the regex change.

Also, fix another bug where second call to `WorkflowReporter.report_completed` was 1) not sending the `version` argument, and 2) sending an undefined `error_message` argument (copy-pasta). I suspect this bug was being hidden by the funky test setup.