Remove an unused initialization argument from `Audit::CatalogToMoab`
I am not sure why this argument was there and not used when the storage location can be grabbed from the CompleteMoab. Possibly used to be used for dependency injection or testing? Instead, grab the storage location from the CompleteMoab since we no longer appear to be using this arg which is mostly getting in the way.