Reporter#moab_detail_csv_list: even fewer queries to run
now instead of querying for the CompleteMoabs for each PreservedObject, leverage the fact that the query already gets most of that info, with the exception of the name of the from_moab_storage_root for each CompleteMoab.  since there are only 10s of storage roots in prod, and since the names are very stable, it's much cheaper to look up each of those names once and cache the result.  we already have the current storage root name in the MoabStorageRoot ivar.  bonus:  easier to get at the column values we want by using #each_row to get a hash, and also no extraneous ActiveRecord objects to instantiate (downside: manual translation of status enum values)