Send audit results to DOR event service
Fixes #1357
Fixes #1351

This pull request sends audit results to the new DOR event service which is provided by dor-services-app. As part of this, I refactored the WorkflowReporter such that the class methods (all that was exposed previously) create an instance of the class, with the instance doing all the work. This preserves the prior contract and allows for less passing around of state which now lives in the instance.