Expect shelve diff endpoint to receive JSON not www-form
This is a case of drift between preservation_catalog (where tests assumed a request body with x-www-form-urlencoded args) and preservation-client (which was sending JSON). Since upstream code is using the client to access this endpoint, the openapi spec (and the tests) should reflect this expectation.

Should resolve https://app.honeybadger.io/projects/50568/faults/59300637