CompleteMoab#migrate_moab should also clear last_version_audit; tighten specs a bit
last_version_audit refers to the last time the moab's version on disk was checked against what the catalog expects the version to be.  this seems like it should be cleared upon re-writing the moab in a different location for the same reasons as last_moab_validation and last_checksum_validation.  test accordingly.

add #migrate_moab testing for a single object move to complete_moab_spec.

add tests to make sure checksum validation job is queued (as opposed to solely testing that the status changed in a way we'd expect to trigger that).