re-enable read-only routes
per https://github.com/sul-dlss/preservation_catalog/issues/1439#issuecomment-600759173 we won't shut down accessioning robots for migration, which means that we'll need to enable the routes that allow accessioning robots to talk to pres cat for their read-only requests.

we'll leave disabled the create and update endpoints, but only the preservation robots should use those (and the preservation robots will still be disabled).