delete storage migration checklist issue template
since we end up doing multiple storage roots at once, it ends up being easier to create a new issue based on one of our previous migration uber tickets.  we use that to coordinate the migration as a whole, instead of spawning a ticket per storage root.  i was initially skittish to get rid of this without having tested the uber ticket approach in a dry run before the first migration, but now that we've successfully used the uber ticket approach for two prod storage migrations, i'm happy to retire this.

much thanks to @ndushay for scaffolding and fleshing out the text for both checklists.