more realistic zip making specs
move from fixtures to mocking as much as possible to make for more realistic tests.  tests still pass with the current code (and are hopefully more illustrative).  the preference for fixtures makes tests less brittle in the face of implementation changes (in particular, this makes it easier to show that an upcoming change passes the same tests the old code passed, adding confidence we're not introducing regressions).