put M2C and C2M audits on monthly schedule
and adjust audit threshold to prevent false alerts.

the alternative (more idiomatic?) native whenever syntax for the job where raw cron syntax was used would be:

every :month, at: 'Jan 15th at 11:00'

the month part of the "at:" param would be disregarded (but without that placeholder month in the "at:" param, '15', '15th', etc, is interpreted as an hour on the first day of the month, not a day of the month).  this appears to be an artifact of whenever's use of chronic (the gem whenever relies on for date parsing).