Remove tests that test initialization (+ side effects)
Fixes #1960

This commit removes tests that are unnecessary as they test behavior that is basic Ruby behavior. This turned out to be a little tricky with the very tangly `CompleteMoabHandler`---its initialization tests covered two tiny private methods that I think we *do* want to continue testing. Instead of elevating those methods to the public API of the CMH class, or testing private methods, I decided to go the route of extracting those two methods to Rails helpers. While I generally am not in favor of leaning on helpers, I felt it was defensible over, say, creating one or two new service classes wrapped around the (small) methods.