Log when total zip part size is less than the moab size
Fixes #1742

This commit adds a new audit result to the catalog-to-archive audit service such that, when the total zip part size for a zipped moab version is less than the size of the moab on disk, this condition is reported. Reporting this information will help us find and remediate replication errors.

As part of this work, as a convenience, add a new `#total_part_size` method to the `ZippedMoabVersion` model.