Move AuditResults model to service class
Fixes #2123

This commit moves what was the AuditResults model---a PORO, not an ActiveRecord model---into `app/services/`, which we admittedly have low and unclear standards around (what is a service object? etc.). But we will be less surprised by finding this class among other junk-drawer-type classes than we will in the directory where its siblings are all ActiveRecord-based models. This is a purely mechanical refactoring.