move Audit::Results::DB_OBJ_ALREADY_EXISTS from audit workflow report codes to honeybadger report codes
this can happen when pres robots alerts pres cat to a moab version that's already been cataloged, as might happen when the update-catalog workflow step sits after having failed, and the moab version is detected by e.g. moab-to-catalog audit before the step is retried.

it's unusual, and worth a glance, but shouldn't wedge preservation or accessioning workflows.