Add a utility to export GIS items from preservation
The gis_exporter utility reads a file of druids and retrieves the latest
version from preservation, and exports the files to a druid directory.
Care is taken to read original files from the data.zip files, and to
ignore metadata files that are created as part of the gis-robot-suite.

References #830