zip endpoint config seeding: tolerate incomplete configuration of new endpoints
when partial settings are provided such that an endpoint entry can't be created, warn and skip, so that deployment can continue.  this allows, for example, setting up config gem env vars that provide access credentials, before the shared_configs override containing the rest of the settings is in place.  this also keeps us from missing the inadvertent omission of a new endpoint if partial information was configured in error.