update gcp endpoint setting and bucket name again
turns out the new endpoint is going to be in the US south region afterall, for parity with IBM, since that region was chosen specifically for geographic diversity compared to the AWS endpionts (which aren't changing in any way at present).

i have an idea, which i'll ticket, for making the code more flexible with respect to hardcoded endpoint names in settings references and the need to coordinate endpoint names across all environments because of the hardcoding.  for now, churn for accuracy and expediency.